requirements_api.txt
# API, Rate Limiting and Caching Dependencies
Django==4.2.16
djangorestframework==3.16.1
django-ratelimit==4.1.0
redis==7.0.1
psycopg2==2.9.9
dj-database-url==3.0.1
django-allauth==65.13.0
gunicorn==20.1.0
whitenoise==6.8.2
